feat(quantities): support decimal quantities app-wide (int -> double)
Storage actors and verifiers need fractional quantities (e.g. 3.5 kg,
0.25 L). Quantities were integers end-to-end, silently rejecting decimals.

- DB: convert quantity/min_quantity/final_quantity/old_quantity/new_quantity
  columns to numeric; recreate 9 RPC functions with ::numeric casts so JSON
  values and locals no longer truncate decimals (usage_count stays integer).
- Models/cubits/repos: quantity fields and params int -> double; fromMap uses
  (x as num).toDouble().
- UI (all roles): decimal input formatter (2 dp max) + numberWithOptions,
  double parsing, and a shared formatQty() that trims trailing zeros so whole
  numbers still render as "3" not "3.0".
- Excel import/export: DoubleCellValue + double parsing.

+++ lib/features/inventory/data/inventory_management_repository.dart
@@ -77,10 +77,10 @@ class InventoryManagementRepository {
   Future<AppResult<void>> createItem({
     required String name,
     required String unit,
-    required int quantity,
+    required double quantity,
     String? sku,
     String? category,
-    int minQuantity = 0,
+    double minQuantity = 0,
     String? description,
     String? notes,
   }) async {
@@ -111,10 +111,10 @@ class InventoryManagementRepository {
     String itemId, {
     required String name,
     required String unit,
-    required int quantity,
+    required double quantity,
     String? sku,
     String? category,
-    int minQuantity = 0,
+    double minQuantity = 0,
     String? description,
     String? notes,
   }) async {
@@ -160,7 +160,7 @@ class InventoryManagementRepository {
   }
 
   Future<AppResult<void>> bulkUpdateQuantities(
-    List<({String itemId, int quantity})> updates,
+    List<({String itemId, double quantity})> updates,
   ) async {
     try {
       final payload = updates
